New BSPC/AAS.

+++ code/bspc/aas_cfg.h
@@ -77,7 +77,7 @@ typedef struct cfg_s
 	float rs_falldamage10;
 	float rs_maxfallheight;
 	float rs_maxjumpfallheight;
-	float rs_allowladders;
+	int rs_allowladders;
 } cfg_t;
 
 extern cfg_t cfg;
